Separating project segments into different files, and putting the main menu into the infinite cycle
