+++ app_components/constants.h
@@ -1,0 +1,10 @@
+#define _USE_MATH_DEFINES // M_PI, M_E, etc.
+#define KEY_UP 65    // Клавиша Вверх
+#define KEY_DOWN 66  // Клавиша Вниз
+#define KEY_ENTER 10 // Клавиша ENTER
+const double sec_per_day = 86400.0; // секунд в дне
+const double au_to_meters = 1.496e11; // метров в 1 астрономической единице
+const double jupiter_mass = 1.898e27; // масса Юпитера в КГ
+const double sun_mass = 1.989e30; // масса Солнца в КГ
+const double jupiter_radius_m = 69911000; // радиус Юпитера в КГ
+float temp_any = -274;
